+++ common/src/main/java/rzk/wirelessredstone/misc/WRProperties.java
@@ -1,0 +1,8 @@
+package rzk.wirelessredstone.misc;
+
+import net.minecraft.state.property.BooleanProperty;
+
+public class WRProperties
+{
+	public static final BooleanProperty LINKED = BooleanProperty.of("linked");
+}