+++ common/src/main/java/rzk/wirelessredstone/registry/ModBlocks.java
@@ -6,6 +6,8 @@ public final class ModBlocks
 {
 	public static Block redstoneTransmitter;
 	public static Block redstoneReceiver;
+	public static Block p2pRedstoneTransmitter;
+	public static Block p2pRedstoneReceiver;
 
 	private ModBlocks() {}
 }