+++ fabric/src/main/java/rzk/wirelessredstone/registry/ModBlocksFabric.java
@@ -6,6 +6,8 @@ import net.minecraft.item.Item;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import rzk.wirelessredstone.WirelessRedstone;
+import rzk.wirelessredstone.block.P2pRedstoneReceiverBlock;
+import rzk.wirelessredstone.block.P2pRedstoneTransmitterBlock;
 import rzk.wirelessredstone.block.RedstoneReceiverBlock;
 import rzk.wirelessredstone.block.RedstoneTransmitterBlock;
 
@@ -17,6 +19,8 @@ public final class ModBlocksFabric
 	{
 		ModBlocks.redstoneTransmitter = registerBlock("redstone_transmitter", new RedstoneTransmitterBlock());
 		ModBlocks.redstoneReceiver = registerBlock("redstone_receiver", new RedstoneReceiverBlock());
+		ModBlocks.p2pRedstoneTransmitter = registerBlock("p2p_redstone_transmitter", new P2pRedstoneTransmitterBlock());
+		ModBlocks.p2pRedstoneReceiver = registerBlock("p2p_redstone_receiver", new P2pRedstoneReceiverBlock());
 	}
 
 	public static Block registerBlockWithoutItem(String name, Block block)