+++ forge/src/main/java/rzk/wirelessredstone/registry/ModBlocksForge.java
@@ -2,6 +2,8 @@ package rzk.wirelessredstone.registry;
 
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegisterEvent;
+import rzk.wirelessredstone.block.P2pRedstoneReceiverBlockWrapper;
+import rzk.wirelessredstone.block.P2pRedstoneTransmitterBlockWrapper;
 import rzk.wirelessredstone.block.RedstoneReceiverBlockWrapper;
 import rzk.wirelessredstone.block.RedstoneTransmitterBlockWrapper;
 import rzk.wirelessredstone.misc.WREvents;
@@ -16,6 +18,8 @@ public final class ModBlocksForge
 		{
 			ModBlocks.redstoneTransmitter = WREvents.register(helper, "redstone_transmitter", new RedstoneTransmitterBlockWrapper());
 			ModBlocks.redstoneReceiver = WREvents.register(helper, "redstone_receiver", new RedstoneReceiverBlockWrapper());
+			ModBlocks.p2pRedstoneTransmitter = WREvents.register(helper, "p2p_redstone_transmitter", new P2pRedstoneTransmitterBlockWrapper());
+			ModBlocks.p2pRedstoneReceiver = WREvents.register(helper, "p2p_redstone_receiver", new P2pRedstoneReceiverBlockWrapper());
 		});
 	}
 }