+++ neoforge/src/main/java/rzk/wirelessredstone/WirelessRedstoneNeo.java
@@ -60,6 +60,8 @@ public class WirelessRedstoneNeo
 				{
 					output.add(ModBlocks.redstoneTransmitter);
 					output.add(ModBlocks.redstoneReceiver);
+					output.add(ModBlocks.p2pRedstoneTransmitter);
+					output.add(ModBlocks.p2pRedstoneReceiver);
 					output.add(ModItems.circuit);
 					output.add(ModItems.frequencyTool);
 					output.add(ModItems.frequencySniffer);