+++ neoforge/src/main/java/rzk/wirelessredstone/registry/ModItemsNeo.java
@@ -20,6 +20,8 @@ public final class ModItemsNeo
 		{
 			WREvents.register(helper, "redstone_transmitter", new BlockItem(ModBlocks.redstoneTransmitter, new Item.Settings()));
 			WREvents.register(helper, "redstone_receiver", new BlockItem(ModBlocks.redstoneReceiver, new Item.Settings()));
+			WREvents.register(helper, "p2p_redstone_transmitter", new BlockItem(ModBlocks.p2pRedstoneTransmitter, new Item.Settings()));
+			WREvents.register(helper, "p2p_redstone_receiver", new BlockItem(ModBlocks.p2pRedstoneReceiver, new Item.Settings()));
 
 			ModItems.circuit = WREvents.register(helper, "circuit", new Item(new Item.Settings()));
 			ModItems.frequencyTool = WREvents.register(helper, "frequency_tool", new FrequencyItem(new Item.Settings()));